+++ src/assets/js/loaders.js
@@ -1,0 +1,4 @@
+export const TEXT_LOADER = `<div class="skeleton skeleton-text"></div>`;
+export const TEXT_BODY_LOADER = `<div class="skeleton skeleton-text skeleton-text__body"></div>`;
+export const IMAGE_LOADER = (size = 200) =>
+  `<img class="skeleton" alt="" height="${size}px" width="${size}px" />`;